+++ illumio-core/data_source_illumio_virtual_services.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 func datasourceIllumioVirtualServices() *schema.Resource {
@@ -287,6 +288,13 @@ func datasourceIllumioVirtualServices() *schema.Resource {
 				ValidateDiagFunc: isStringABoolean(),
 				Description:      "Include Virtual Service usage flags",
 			},
+			"match_type": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Default:      PARTIAL_MATCH,
+				ValidateFunc: validation.StringInSlice([]string{PARTIAL_MATCH, EXACT_MATCH}, true),
+				Description:  `Indicates whether to return all partially-matching names or only exact matches. Allowed values are "partial" and "exact". Default value: "partial"`,
+			},
 		},
 	}
 }
@@ -352,6 +360,13 @@ func dataSourceIllumioVirtualServicesRead(ctx context.Context, d *schema.Resourc
 	}
 
 	for _, child := range data.Children() {
+		// if exact matching is enabled, skip the object if it's a partial match
+		if d.Get("match_type").(string) == EXACT_MATCH {
+			if !isExactMatch("name", d, child) {
+				continue
+			}
+		}
+
 		m := extractMap(child, keys)
 
 		key := "service"